+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Adds app owned WNFS
 - Separates initialize into app and permissionedApp entrypoints
 - Make email optional at registration
+- Add reset option to bootstrapFileSystem
 
 ### v0.32.0